api/requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
Werkzeug==0.16.1
Flask-Cors==3.0.8
google-cloud-datastore==1.10.0
#google-auth==1.10.0
#requests==2.22.0
mgrs==1.3.6
pyyaml==5.3